requirements.txt
defusedxml==0.7.1
fonttools==4.53.0
fpdf2==2.7.9
pillow==10.3.0
